// src/api/or-planner/models/ScheduledOperation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OR Planner API
 * Operating Room Planner API
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: filip.chromek@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ScheduledOperation
 */
export interface ScheduledOperation {
    /**
     * 
     * @type {string}
     * @memberof ScheduledOperation
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof ScheduledOperation
     */
    patientId: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof ScheduledOperation
     */
    staffIds?: Array<string>;
    /**
     * 
     * @type {string}
     * @memberof ScheduledOperation
     */
    procedureType: string;
    /**
     * 
     * @type {number}
     * @memberof ScheduledOperation
     */
    priority?: number;
    /**
     * 
     * @type {Date}
     * @memberof ScheduledOperation
     */
    scheduledStart: Date;
    /**
     * 
     * @type {number}
     * @memberof ScheduledOperation
     */
    durationMinutes: number;
    /**
     * 
     * @type {string}
     * @memberof ScheduledOperation
     */
    status?: ScheduledOperationStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof ScheduledOperation
     */
    notes?: string;
}


/**
 * @export
 */
export const ScheduledOperationStatusEnum = {
    Planned: 'planned',
    InProgress: 'in-progress',
    Completed: 'completed',
    Cancelled: 'cancelled'
} as const;
export type ScheduledOperationStatusEnum = typeof ScheduledOperationStatusEnum[keyof typeof ScheduledOperationStatusEnum];


/**
 * Check if a given object implements the ScheduledOperation interface.
 */
export function instanceOfScheduledOperation(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "patientId" in value;
    isInstance = isInstance && "procedureType" in value;
    isInstance = isInstance && "scheduledStart" in value;
    isInstance = isInstance && "durationMinutes" in value;

    return isInstance;
}

export function ScheduledOperationFromJSON(json: any): ScheduledOperation {
    return ScheduledOperationFromJSONTyped(json, false);
}

export function ScheduledOperationFromJSONTyped(json: any, ignoreDiscriminator: boolean): ScheduledOperation {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'patientId': json['patientId'],
        'staffIds': !exists(json, 'staffIds') ? undefined : json['staffIds'],
        'procedureType': json['procedureType'],
        'priority': !exists(json, 'priority') ? undefined : json['priority'],
        'scheduledStart': (new Date(json['scheduledStart'])),
        'durationMinutes': json['durationMinutes'],
        'status': !exists(json, 'status') ? undefined : json['status'],
        'notes': !exists(json, 'notes') ? undefined : json['notes'],
    };
}

export function ScheduledOperationToJSON(value?: ScheduledOperation | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'patientId': value.patientId,
        'staffIds': value.staffIds,
        'procedureType': value.procedureType,
        'priority': value.priority,
        'scheduledStart': (value.scheduledStart.toISOString()),
        'durationMinutes': value.durationMinutes,
        'status': value.status,
        'notes': value.notes,
    };
}